+++ R/utils.R
@@ -8,14 +8,14 @@ gettime <- function()
     else
         return( as.double( base::Sys.time() ) )
     }
-    
+
 
 #   projectiveMatrix()
-#    
+#
 #   .matrix     invertible matrix, for example a 3x3 matrix with columns the tristimulus coordinates of RGB primaries
 #   .unit       non-zero vector.  For example the tristimulus coordinates of white.
 #
-#   return      square matrix  B, so that   
+#   return      square matrix  B, so that
 #               B = matrix  %*%  diag(a)  <=>   each column of B is a multiple of the corresponding column in .matrix
 #               B %*% 1  =  .unit.      (1 is the vector of all 1s)
 #
@@ -30,30 +30,30 @@ gettime <- function()
 projectiveMatrix  <-  function( .matrix, .unit )
     {
     a   = try( solve( .matrix, .unit ), silent=TRUE )
-    
+
     if( ! is.numeric(a) ) return(NULL)
-    
+
     ran = range( abs(a) )   #; print(ran)
-    
+
     if( ran[1] <= 1.e-6 * ran[2] ) return(NULL)
-    
+
     return( .matrix  %*%  diag(a) )
     }
-    
-    
-#   return list with A and B    
+
+
+#   return list with A and B
 ABfromHC <- function( H, C )
     {
     theta   = H * pi/50
     list( A = C*cos(theta), B = C*sin(theta) )
     }
-    
+
 HCfromAB <- function( A, B )
     {
     theta   = atan2( B, A )
     list( H=(theta * 50/pi) %% 100, C = sqrt( A^2 + B^2 ) )
-    }    
-    
+    }
+
 hypot<-function(a, b){
 # sqrt(a^2 + b^2) without under/overflow.  Author: Jose Gama **/
 # http://www.java2s.com/Tutorial/Java/0120__Development/sqrta2b2withoutunderoverflow.htm
@@ -66,8 +66,8 @@ if (abs(a) > abs(b)) {
          r <- abs(b)*sqrt(1+r^2)
       }
 r
-}    
-    
+}
+
 ###########     argument processing     ##############
 #
 #   A   a non-empty numeric Nx3 matrix, or something that can be converted to be one
@@ -82,25 +82,26 @@ r
 #   This is intended to check user-supplied A, so there is a lot of checking.
 #
 prepareNx3  <-  function( A, M=3 )
-    {    
+    {
     ok  = is.numeric(A) &&  0<length(A)  &&  (length(dim(A))<=2)
-    
+
     ok  = ok  &&  ifelse( is.matrix(A), ncol(A)==M, ((length(A) %% M)==0)  )
-    
+
     if( ! ok )
         {
         mess    = substr( as.character(A)[1], 1, 20 )
-        
-        Aname = deparse(substitute(A))        
-             
-        log.string( ERROR, "Argument '%s' must be a non-empty numeric Nx%d matrix (with N>0). %s='%s...'", 
-                                    Aname, M, Aname, mess )
+
+        Aname = deparse(substitute(A))
+
+        #   notice hack to make log_level() print name of parent function, and *NOT* prepareNx3()
+        log_level( ERROR, "Argument '%s' must be a non-empty numeric Nx%d matrix (with N>0). %s='%s...'",
+                                    Aname, M, Aname, mess, .topcall=sys.call(-1L) )
         return(NULL)
         }
-    
+
     if( ! is.matrix(A) )
         A = matrix( A, ncol=M, byrow=TRUE )
-        
+
     return( A )
     }
 
@@ -115,18 +116,17 @@ prepareNx3_old  <-  function( A, M=3 )
         #print( sys.frames() )
         mess    = substr( as.character(A)[1], 1, 10 )
         #arglist = list( ERROR, "A must be a non-empty numeric Nx3 matrix (with N>0). A='%s...'", mess )
-        #do.call( log.string, arglist, envir=parent.frame(n=3) )
-        #myfun   = log.string
+        #do.call( log_level, arglist, envir=parent.frame(n=3) )
+        #myfun   = log_level
         #environment(myfun) = parent.frame(3)
-        log.string( ERROR, "Argument A must be a non-empty numeric Nx%d matrix (with N>0). A='%s...'", M, mess )
+        log_level( ERROR, "Argument A must be a non-empty numeric Nx%d matrix (with N>0). A='%s...'", M, mess )
         return(NULL)
         }
-    
+
     if( is.null(dim(A)) )
         A = matrix( A, ncol=M, byrow=TRUE )
     else if( ncol(A) != M )
         A = t(A)
-        
+
     return( A )
     }
-